three-spin ising basis reduction
this constructs all the basis sectors corresponding to the maximal set of mutually commuting operators in Three-Spin Ising chain, with or without the further reduction with momentum block for periodic chains. 

Observation: Although the ground state is analytically known to be in the (D1, D2, D3) = (+1, +1, +1) sector, all the other sectors are degenerate at h=0. [This causes issues to be undestood thoroughly!!]